Now the Wyvern interpreter implements mathematical integers!

+++ tools/src/wyvern/stdlib/support/Stdio.java
@@ -1,11 +1,13 @@
 package wyvern.stdlib.support;
 
+import java.math.BigInteger;
+
 public class Stdio {
 	public static Stdio stdio = new Stdio();
 	public void print(String text) {
 		java.lang.System.out.print(text);
 	}
-	public void printInt(int n) {
+	public void printInt(BigInteger n) {
 		java.lang.System.out.print(n);
 	}
 }
